+++ bundle/src/main/java/com/adobe/acs/commons/redirects/models/RedirectRule.java
@@ -51,6 +51,7 @@ public class RedirectRule {
     public static final String TARGET_PROPERTY_NAME = "target";
     public static final String STATUS_CODE_PROPERTY_NAME = "statusCode";
     public static final String UNTIL_DATE_PROPERTY_NAME = "untilDate";
+    public static final String EFFECTIVE_FROM_PROPERTY_NAME = "effectiveFrom";
     public static final String NOTE_PROPERTY_NAME = "note";
     public static final String CONTEXT_PREFIX_IGNORED_PROPERTY_NAME = "contextPrefixIgnored";
     public static final String CREATED_PROPERTY_NAME = "jcr:created";
@@ -70,6 +71,9 @@ public class RedirectRule {
 
     @ValueMapValue
     private Calendar untilDate;
+
+    @ValueMapValue
+    private Calendar effectiveFrom;
 
     @ValueMapValue
     private String note;
@@ -154,16 +158,12 @@ public class RedirectRule {
         return modified;
     }
 
-    /**
-     * This is ugly, but needed to format untilDate in HTL in the redirect-row component.
-     * <p>
-     * AEM as a Cloud Service supports formatting java.time.Instant (SLING-10651), but
-     * classic AEMs 6.4 and 6.5 only support formatting java.util.Date and java.util.Calendar
-     *
-     * @return java.util.Calendar representation of untilDate
-     */
     public Calendar getUntilDate() {
         return untilDate;
+    }
+
+    public Calendar getEffectiveFrom() {
+        return effectiveFrom;
     }
 
     public String[] getTagIds() {
@@ -190,9 +190,9 @@ public class RedirectRule {
 
     @Override
     public String toString() {
-        return String.format("RedirectRule{source='%s', target='%s', statusCode=%s, untilDate=%s, note=%s, "
+        return String.format("RedirectRule{source='%s', target='%s', statusCode=%s, untilDate=%s, effectiveFrom=%s, note=%s, "
                         + "contextPrefixIgnored=%s, tags=%s, created=%s, createdBy=%s, modified=%s, modifiedBy=%s}",
-                source, target, statusCode, untilDate, note, contextPrefixIgnored,
+                source, target, statusCode, untilDate, effectiveFrom, note, contextPrefixIgnored,
                 Arrays.toString(tagIds), created, createdBy, modified, modifiedBy);
     }
 
@@ -235,5 +235,38 @@ public class RedirectRule {
             ptrn = null;
         }
         return ptrn;
+    }
+
+    /**
+     * ----[effectiveFrom]---[now]---[untilDate]--->
+
+     *
+     * @return
+     */
+    public RedirectState getState(){
+        boolean expired = untilDate != null && untilDate.before(Calendar.getInstance());
+        boolean pending = effectiveFrom != null && Calendar.getInstance().before(effectiveFrom);;
+        boolean invalid = effectiveFrom != null && untilDate != null && effectiveFrom.after(untilDate);
+
+        if (invalid){
+            return RedirectState.INVALID;
+        } else if (expired){
+            return RedirectState.EXPIRED;
+        } else if (pending){
+            return RedirectState.PENDING;
+        } else {
+            return RedirectState.ACTIVE;
+        }
+    }
+
+    /**
+     * @return whether the redirect is published
+     */
+    public boolean isPublished(){
+        Calendar lastReplicated = resource.getParent().getValueMap().get("cq:lastReplicated", Calendar.class);
+        boolean isPublished = lastReplicated != null;
+        boolean modifiedAfterPublication = isPublished
+                && ((modified != null && modified.after(lastReplicated)) || (created != null && created.after(lastReplicated)));
+        return isPublished && !modifiedAfterPublication;
     }
 }